ENHANCE: Separate audio file request to an individual module
